+++ Assets/@Model/Obstacle/Scripts/Obs_Bounce.cs
@@ -1,0 +1,33 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Bounce : MonoBehaviour
+{
+    public float force = new float(); //Force 10000f
+	public float stunTime = new float();
+	private Vector3 hitDir;
+
+	void OnCollisionEnter(Collision collision)
+	{
+		foreach (ContactPoint contact in collision.contacts)
+		{
+			Debug.DrawRay(contact.point, contact.normal, Color.white);
+			if (collision.gameObject.tag == "Player")
+			{
+				hitDir = contact.normal;
+				collision.gameObject.GetComponent<PlayerController>().HitPlayer(-hitDir * force, stunTime);
+				return;
+			}
+		}
+		/*if (collision.relativeVelocity.magnitude > 2)
+		{
+			if (collision.gameObject.tag == "Player")
+			{
+				//Debug.Log("Hit");
+				collision.gameObject.GetComponent<CharacterControls>().HitPlayer(-hitDir*force, stunTime);
+			}
+			//audioSource.Play();
+		}*/
+	}
+}
