+++ Assets/@Model/Obstacle/Scripts/Obs_Rotator.cs.meta
